refactor: Reservation 도메인 통합 리팩토링 myreservation, reservation-status 도메인을 단일 reservation 도메인으로 통합하여 예약 관련 코드 중복을 제거하고 응집도를 높임 🔧 주요 개선사항: - 중복 제거: StateType, StateBadge 등 중복 코드 완전 제거 - 타입 통합: 3개 중복 타입 파일 → 1개 통합 types.ts - API 표준화: 사용자/호스트 관점별 API 함수 통합
