+++ src/domain/reservation/components/ReservationStatusClient/formatDateString.ts
@@ -1,0 +1,3 @@
+export const formatDateString = (date: Date): string => {
+  return `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, '0')}-${String(date.getDate()).padStart(2, '0')}`;
+};
