+++ src/domain/reservation/components/StateBadge/type.ts
